api/requirements.txt
alembic==1.11.2
aniso8601==9.0.1
astroid==2.15.6
blinker==1.6.2
cffi==1.15.1
click==8.1.6
cryptography==40.0.2
dill==0.3.7
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mccabe==0.7.0
msgpack==1.0.5
packaging==23.1
platformdirs==3.10.0
pycparser==2.21
PyJWT==2.8.0
pylint==2.17.5
pynvim==0.4.3
python-dotenv==0.21.1
python-dotenv-vault==0.6.1
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.19
tomlkit==0.12.1
typing_extensions==4.7.1
Werkzeug==2.3.6
wrapt==1.15.0
